[feat/#44] 상세 내역 모집 domain model, viewModel
