+++ app/src/main/java/com/poti/android/domain/model/history/ParticipantManage.kt
@@ -2,8 +2,8 @@ package com.poti.android.domain.model.history
 
 import com.poti.android.domain.type.ParticipantStatusType
 
-data class ParticipantManageDetail(
-    val recruitId: Int,
+data class RecruiterDetail(
+    val partyId: Long,
     val artistInfo: ArtistInfo,
     val progressInfo: ProgressInfo,
     val participantInfoList: List<ParticipantInfo>,
@@ -23,7 +23,7 @@ data class ProgressInfo(
 )
 
 data class ParticipantInfo(
-    val userId: Int,
+    val userId: Long,
     val memberNames: String,
     val participantState: ParticipantStatusType,
     val userInfo: String,
