+++ app/src/main/java/com/poti/android/presentation/history/navigation/HistoryNavigation.kt
@@ -24,7 +24,7 @@ sealed interface HistoryRoute : Route {
     data object RecruiterDetail : HistoryRoute
 
     @Serializable
-    data object ParticipantManage : HistoryRoute
+    data class ParticipantManage(val recruitId: Long) : HistoryRoute
 }
 
 fun NavController.navigateToHistoryList() {
@@ -39,12 +39,13 @@ fun NavController.navigateToRecruiterDetail() {
     navigate(HistoryRoute.RecruiterDetail)
 }
 
-fun NavController.navigateToParticipantManage() {
-    navigate(HistoryRoute.ParticipantManage)
+fun NavController.navigateToParticipantManage(recruitId: Long) {
+    navigate(HistoryRoute.ParticipantManage(recruitId))
 }
 
 fun NavGraphBuilder.historyNavGraph(
     paddingValues: PaddingValues,
+    navController: NavController
 ) {
     composable<HistoryRoute.HistoryList> {
         HistoryListRoute(modifier = Modifier.padding(paddingValues))
@@ -53,9 +54,15 @@ fun NavGraphBuilder.historyNavGraph(
         ParticipantDetailRoute(modifier = Modifier.padding(paddingValues))
     }
     composable<HistoryRoute.RecruiterDetail> {
-        RecruiterDetailRoute(modifier = Modifier.padding(paddingValues))
+        RecruiterDetailRoute(
+            modifier = Modifier.padding(paddingValues),
+            // TODO: [천민재] 마이페이지 모집 내역으로 이동
+            onBackClick = {},
+            // TODO: [천민재] 분철팟 상세 페이지로 이동
+            onDetailClick = {},
+            onParticipantManageDetailClick = navController::navigateToParticipantManage)
     }
     composable<HistoryRoute.ParticipantManage> {
-        ParticipantManageRoute(modifier = Modifier.padding(paddingValues), onBackClick = {} )
+        ParticipantManageRoute(modifier = Modifier.padding(paddingValues))
     }
 }
