+++ app/src/main/java/com/poti/android/presentation/main/PotiNavHost.kt
@@ -34,6 +34,7 @@ fun PotiNavHost(
             paddingValues = paddingValues,
         )
         historyNavGraph(
+            navController = navigator.navController,
             paddingValues = paddingValues,
         )
         myPageNavGraph(
